src/api-wrapper/requirements.txt
fastapi
fastapi-camelcase
python-multipart
python-dotenv
aiofiles
